Fix concurrency issue in send_flow_control_msg()
It can be reached by all calling thrads of the transports send and receive
functions. With DRBD this is at lease the receiver thread, an "areceiver"
thread and probably more thrads sending on the control socket.

dtr_recv_pages(), dtr_recv()
  dtr_receive_rx_desc()
      dtr_send_flow_control_msg()

dtr_recycle_rx_desc()
  dtr_flow_control()
      dtr_send_flow_control_msg()

The interesting thing is that dtr_send_flow_control_msg() determins
the number of new tx_descs by calculating posted - known, and then
increasing known by the difference. ... If that happens in multiple
threads in parallel, we might find the difference twice and
adding it twice to known...

Fix that by introducing a spin_lock protecting this area.