drbd: Apply auto-promote-timeout also for read-only openers
So far, when a newly created resource was opened for read & write
we waited for auto-promote-timeout if a peer with consistent
data shows up.

If one tied to open it read-only we failed immediately.

Make that more consistent by also waiting for auto-promote-timeout
if a peer with valid data show up in case it gets opened for
read only access.