drbd: Hold a reference while waiting for valid data in drbd_open()
Since auto_promote was introduced there was this race:
One process tries to open a drbd device and sleeps in drbd_open() in
try_to_promote() or in wait_event_i...(...,ro_open_cond(device)).
An other process comes by and removes the minor/device via the netlink
interface.

Fix that by holding a reference on the device.