drbd: Initialize twopc_type on initiator node
What happened here is that a P_TWOPC_YES got interpreted as for
a TWOPC_STATE_CHANGE instead of TWOPC_RESIZE.

In the logfile you would see the problem like this:

drbd r0/0 drbd7: drbd_bm_resize called with capacity == 68700602544
drbd r0/0 drbd7: Not enough space for bitmap: 8587575318 > 529728
drbd r0/0 drbd7: BM resizing failed. Leaving size unchanged
drbd r0/0 drbd7: size = 1024 MB (1048600 KB)

The fix obviously is to initialize to twopc_type when starting the
transaction on the initiator node.

The bug is only visible if there is a diskless node, and the reply
of the diskless node gets in last. As the diskless node does not
want to limit the new size it suggests DRBD_MAX_SECTOR.

That is 32 TByte.

That should be 1PByte and the DRBD_MAX_SECTOR_FLEX define. Fixing
that in the same commit.