drbd: Timeout for downing a device (udev workaround)
In case you demote a device, and immediately afterwards
you try to down it, it might be that udev opened it in read-only
mode to investigate it after it became secondary. We failed the
down operation if there are read-only openers.

For the user it looks like the down command is unreliable.

Fix that by waiting up to a second for such read-only openers
to go away.