drbd: don't ignore fencing policy just because we don't see a primary
The whole "from_consistent" mechanism is intended for the case when
a secondary looses the connection to its primary, and it does not know
if there might be an other path to the primary.

While it does not know it goes temporarily to D_CONSISTENT.

This whole machinery should not trigger when a node comes up and
is consistent since it obeys the fencing policies.