drbd: Always have quorum if everybody away is D_OUTDATED
This is an optimization that allows a kind of "last man standing".
As long as you remove nodes gracefully (i.e. a clean down or a clean
disconnect), the primary will continue to work even if it is
the last node.
That works by marking the leaving nodes as D_OUTDATED, and the
primary node knows that they are D_OUTDATED, and therefor can
not be promoted to primary, even if they are the majority.

Only if we loose a node hard, we need to assume that this node
joins the gang of D_OUTDATED nodes, if they have majority
we have lost quorum and need to freeze.