change link to a more static version
Recently the sales department likes to play around with the web page,
which breaks various links all the time. I added one link (in this
change) that forwards to the "human readable" web presence, and another
(http://links.linbit.com/sources) that links to the root of the actual
download directories. The second one should be used for various scripts
that want to fetch the source code (e.g. Alpine PKGBUILDs).