drbd: Fix late application of exposed-data-uuid update
This is a fix for 8762c2d52
'drbd: Do not wait for cluster state changes to complete in the receiver'

The issue is rather awkward. The previous code, could lead to such
an update being delayed ways to long, effectively setting the exposed
data uuid back in time. Which can cause a lot of confusion.