Fixup compat for ib_get_dma_mr() removal
The replacedment for ib_get_dma_mr() is device->get_dma_mr(), but
it missed to do:

  mr->device = pd->device;
  mr->pd = pd;
  mr->uobject = NULL;
  mr->need_inval = false;

that resulted in a crash when trying to do ib_dereg_mr().

I decided against hiding this backward compatibility code, by doing
the missing initiazation of mr, but rather have the new upstream
way clearly visible in the code. That involves #ifdef'in the old
code in, but that makes it clear. I think that is better in this
case. (Though generally I like to hide the backward compat cruft)