Remove dtr_path->ib_device
Remove everything that relates to a RDMA/IB connection from dtr_path
and have it in dtr_cm. This patch is the first step.

The reason is that when establishing a connection the active and the
passive side can not arbitrary call __dtr_unprepare_path(), since
the other one might just started to use it...