Fix keeping wrong cm in dtr_path_prepare()
When there is a race between a connect() and accept(), i.e. we succeed to
get an (active) connection and at the same time we get (passively) a
connection in we need to keep one and to close the other one.

The previous code without the cmpxchg() was all wrong. It updated the
path->cm and then returned to free the feshly assigned cm!