fix rcu use of path->cm
a
rcu_read_lock()
cm = rcu_dereference(path->cm)
...
rcu_read_unlock()

only guarantees that the cm does not go away before the rcu_read_unlock().
It does _NOT_ guarantee that the members of the cm get destroyed. For
such a kind of guarantee you need to get a reference on the cm.
...the rcu lock allows to do that...