Keep a reference to the cm from every rx_desc
We have a pointer from every rx_desc to the cm, therefore that should hold
a reference on the cm as well.

There where many ways how we ended up in dereferencing this rx_desc->cm
pointer after the cm was destroyed.
Of course removing the rx_descs needs to move the the __dtr_disconnect_path()
function from the destructor, since we can not drop the references from the
destructor...