drbd: discard/write-same: fix potential distributed deadlock
Currently, the receiver side handles "batch bios" like discard or write-same
synchronously in the receiver context.
Which may now have to block on the activity log.
If there are no free slots available, it waits for P_PEER_ACKs
P_PEER_ACK are sent only once the primary knows things are on stable storage on
the peers, which it only knows once the corresponding P_BARRIER_ACK arrives,
which is only sent as a response to a P_BARRIER, which would need to be received
by the receiver which is currently blocking :-(

This fix is actually more like a workaround: artificially shorten the "current
epoch", and put any "batch bio" request as a first request into a new epoch.
That way, even if the receiver needs to block on that batch request, we know
that whatever was still pending, will "soon" be P_BARRIER_ACKed, and
subsequently P_PEER_ACKed, and we will make progress.

This may induce additional latency to zero-out/trim/discard or write-same,
but at least it won't block.

To-be-done: to improve e.g. fstrim performance, we'd need to handle batch
requests asynchronously on the receiving side as well, and/or introduce an
"unsolicited P_BARRIER_ACK" to trigger P_PEER_ACKs we are waiting for.