drbd: Fix initial resync
The purpose of commit 7842895 was to inform the peers about the state
transition D_INCONSISTENT -> D_UP_TO_DATE. When a node is forced into
primary state, to trigger the initial resync we observe the same
transition on the disk state. That needs to be an exception to not
break the initial resync.