drbd: Fix attach of first disk
In case you detached all disks from a resource (on all nodes! Only possible
if there is no primary) and try to re-attach one, the attach of the first
one should not go though the negotiating state. Since there is no other
disk to do a resync from.

The first attach needs to get the disk state purly from the meta-data.

The expression was wrong. It only made sure that the peer-disks are visible,
it also needs to ensure that the peer is not diskless to consider the
peer as "with disk".