drbd: consider a node as unstable after complete handshake to a R_PRIMARY
The drbd_device_stable() function returns if the data-set of a node is
considered as unstable.

A primary is stable since it is authoritative.
Unstable are neighbors of a primary and resync target nodes.
Nodes further away from a primary are stable!

During handshake (initial state, UUIDs get exchanged, ...) cstate is already
connected but repl_state is still < L_ESTABLISHED. Rely on repl_state
instead of cstate so that it returns the view of the world before this
connection is established.

All that change is necessary to get rid of false
"Ignore Split-Brain, for now, at least one side unstable".
See commit a99fa8c998ad2759384245ad168744baaa1ed150