drbd: Fix sending bitmap UUIDs when node_id > 3
The for_each_set_bit() macro takes the number of bits, not bytes.
The smallest possible patch is:

-       for_each_set_bit(i, (unsigned long *)&bitmap_uuids_mask, sizeof(bitmap_uuids_mask))
+       for_each_set_bit(i, (unsigned long *)&bitmap_uuids_mask, DRBD_NODE_ID_MAX)

decided to remove the whole for_each_set_bit() loop, since it's body
can be trivially folded into the previous loop.

The symptoms of this bug are that you will experience split-brains,
when a regular resync is expected. The bug triggers only for the nodes
with a node_id > 3.