fixup! compat: Use refcount_t 'atomic' type to implement structref
The whole point of using the atomic_t there was to get around the
check in kref_get() to not resurrect an dead object. At this point
we are resurrecting an drbd_request object by purpose. I.e. not
use refcount_inc(), which again has a check to warn about object
resurrection, but refcount_set().