drbd: silence false-positive compiler warning
Gcc is right that in case idx == -1 then old_net does not get
initialized. But what it does not comprehend is that in the line where gcc
complaines about old_net being read, it is in (idx == -1 || old_net == ...)
So since idx == -1 the right sinde of the || will not get evaluared
anyways. Whatever. Make the compiler happy.