drbd: Fix race between multi-volume autopromte/autodemote
In a two-volume resource, one doing an autopromote of volume 0. While
that does it's two phase commit, a read-only opener closes volume 1.

What happened so far is, that the opener has not yet increased rw_cnt,
and the closer sees rw_cnt == 0 and role == R_PRIMARY, therefore it
demotes it to secondary.

Fix that by only considering to demote it if I am a read-write opener
myself. Also increase the rw_cnt before doing anything, and giving it
up in case auto-promote failed.

The side effect of that is that rw_cnt might be positive although
the current role is R_SECONDARY. That is not a problem for the two
readers of rw_cnt (in drbd_state.c and drbd_nl.c)

The open_release mutex is necessary to protect the testing if
open_rw_cnt == 0 and the actual demote against an other thread
increasing open_rw_cnt again.