drbd: Ensure to process all writes before becoming secondary
Especially in systems with auto promote (and that implies also
auto demote), a resource may get demoted very soon after a write
request.

If that write request happens to be handled by the submitter
workqueue, then it happened that we where able to change the
role to secondary _before_ the submitter handled that write
request.