drbd: In case first connection to diskless/primary also consider to downgrade own disk state
if the current UUID does not match.

Implement it on both sides of the connection. That way we avoid that the
diskless primary sees the newly joining node for a short time as D_UP_TO_DATE.

This bug hat fatal consequences. I.e. it can cause the primary to read-back
partially stale data from the newly joined node. I.e. it presents corrupted
data to the reader.