drbd: switch to use call_rcu() drbd_adm_del_minor()
it turns out that the synchronize_rcu() takes considerable amount of
time when deleting many devices.

The idea is to drop the last kref out of the call_rcu() callback.

That means that on the read-sides, the assumpltion that everything
that can be found in an RCU'd list or idr still has a kref ref
of at least one is still true.

The consequence is that the _destroy_ callback for kref_put may
not sleep. (Which is a good idea anyways)