drbd: use call_rcu() while deleting paths
it turns out that the synchronize_rcu() takes considerable amount of
time when deleting many paths.

The idea is to drop the last kref out of the call_rcu() callback.

That means that on the read-sides, the assumpltion that everything
that can be found in an RCU'd list or idr still has a kref ref
of at least one is still true.