drbd: Do not unregister objects twice
At least with resources it was possible to remote it twice, by calling
drbdsetup del-resource in parallel.

Both threads find the resource get a kref on it and are going to unregister
it. Though the list_del_rcu() may not be called twice!

Use the unregister bit to avoid that. Do the same for connections
and minors out of symmetry.