drbd: fix misaligned bios
with the drbd-9.0.9 release a bug in the compat wrappers caused all our
BIOs from our private biosets to be missaligned in main memory.

Nobody noticed on the x86 platform, but it is a problem for our users on
PPC64LE or on ARM.

With that commit I also remove the not used drbd/compat directory.