drbd: Fail a two-phase connect when peer primary and a local ro opener present
The problem was that before doing phase one of the two-phase commit we
do not know that the peer is primary.
We know that after phase one. Up to now the code failed to abort
the transaction at that point, but sent a commit to the other nodes
and failed then in commiting the state change locally!