drbd: Make sure to return IO errors as soon as we lost quorum
In order to get the quorum behaviour right it is important to
return an IO error (if the user which to get IO errors upon loss
of quorum) with completing the first request after we found
out that we lost quorum.

This is the first step, see the next commit for a complete
description.