drbd: Improve quorum implementation
the problem:
When a primary node lost quorum, the other partition got
quorum and the service was started in the other partition,
the previous primary was not able to joint the new masters
because the data generation code reported a split brain.

Let's look at the possible failure scenarious to understand this
change:

The nodes are named: A, B, C whereas for the following A is primary and
some application is accessing files on the mounted file system.

B or C fails:

A - C or A - B still have quorum application continues to run.

C and B fail:

A needs to freeze IO. Important: it needs to freeze or fail any open
write request. As DRBD does the local disk write concurrently with sending
the data to its peers, it might have written the block to disk at the
moment the "lost quorum" condition is detected.
I.e. it might be that DRBD wrote it successfully to the local disk, but
it returns an IO error.

Now the application might shut down due to the IO error, and unmount
the file system and demote the node to Secondary.

On the side of DRBD we need to keep the current-UUID as it is. We can
not create a new one since A does not have the quorum. (A need to assume
that C and B formed a partition with quorum!)

In case C and B stayed secondary and the nodes connect again, we need to
do a resync. We will forward B and C to the A's version of the data.

The implementation will be based on a new meta-data flag. Probably it is
going to be called "primary-lost-quorum".

In case C and B form a cluster with quorum and one of the promotes, it
creates a new current UUID. If A joins later, the "completed-with-error-write"
will be rolled back by the resync from the new current UUID.

the "primary-lost-quorum" flag is only relevant if the joining nodes have
the same current UUID.

NB, "primary-lost-quorum" overrules the "crashed_primary" flag.

The "primary-lost-quorum" flag will be set if a primary node loses quorum.
The flag will be persisted in the meta-data area.

The flag gets cleared in the moment it re-gains quorum since a node joins
it. In that very same moment, it creates a new current UUID. That is necessary
to have the correct resync in place for later joining nodes.

Apart from these in changes to the driver code, it will also export the
quorum state to user-space, so that the pacemaker resource agent can
take this into account when calculating the "master score".