drbd: Fixup for the late new-current-uuid logic
When a peer is lost, and no block is in flight, we can delay
creating the new-current-uuid until the next write arrives.

But if a write was in flight, that means we created the new
data-set in that moment. So create that, send it and persist
it locally even before completing these requests that were
in flight to the just lost peer.

BTW, the NEW_CUR_UUID bit is already set with the state
transition with that the peer goes away.

BUT! If we are loosing quorum in the same moment, that
meands we will complete the pending requests with errors,
that means delay the creation of the new current uuid
further.