drbd: Make sure to not see a temporary fail when downing an resource
The down command needs to demote a primary to secodnary first, that in
usually triggers a scan run of udev. -> Some udev scan process opens
the device for read only.
In the meantime the down operation has demoted the device, disconnected
any peers and finally wants to detach the disk. That may fails since DRBD
wants to serve data to the read-only opener.

Fix: While doing a down operation have the DOWN_IN_PROGRESS bit set on
the resource. If some process tries to open while this bit is set we
DRBD declines that immediately with -EAGAIN.