drbd: Set SO_KEEPALIVE on data sockets
For nearly 20 years this was not necessary, since for most of the time DRBD was
a two node technology.

In case there are 3 nodes with DRBD connections up. I.e. two TCP connections
each.  One node is primary for a long time. It writes data, that gets
replicated to the two secondaries. That generates traffic on both TCP
connections from the primary node to both secondaries.

The DRBD connection between the two secondaries is idle. On one of the two
TCP connections (the "control" socket) DRBD exchanges every 6 seconds a pair
of small packets to make sure the peer is still alive.
But the other TCP connection (the "data" socket) is really idle all the time.

There is some network equipment sitting between the nodes that does connection
tracking ("firewall"). -- The connection tracking timeout is smaller than the
time the primary node does its job. The existence of that long idle TCP
connection is dropped from the tracking table.

In the moment one of the two secondaries wants to promote, it sends the request
to promote itself over the data socket. Now, the firewall in between sees a
packet it has no connection for in its tracking table. It drops it. (A bit nicer
would be to send back a TCP-RST packet, then the host OS would tell DRBD that
the TCP connection broke down; but all that does not happen)

DRBD sends its DRBD-pings over control socket as it does not get a reply
to its request for promoting, and the DRBD-ping-ack comes back... so the peer
must be alive...

...better use SO_KEEPALIVE on data sockets.