drbd: Clear the discard-my-data flag when the the first sync-handshake happens
So far the discard-my-data flag was cleared in the conn_disconnect()
function. That functions gets called when we error while trying to
establish a connection.

In case the user has a network latency that causes the connect
process to timeout a few times before it gets established, the user
got the impression the discard-my-data simply does not work.

Circumvent that by clearing it only when when DRBD does the first
sync-handshake for a volume.