drbd: Abort resync if a sync-source node becomes weakly connected
If the sync source becomes weakly connected it became outdated
at the same time. It might be that The sync-target gets updates
from the primary that would be overwritten by the resync from
the outdated node!

The resync needs to be aborted before the formre resync-target
can process the first write request from the primary.

TODO: Make sure the nodes does not accept any outstanding
RS_reply packtes