drbd: Resolve a possible splitbrain "after unstable" automatically
In case of 3 nodes with only two connections, that make them a chain,
one end is weakly connected if the other end is primary.

A -- B -- C (where A is primary)

When the primary writes data, a bitmap build up to the weakly connected
node. When the primary disconnects from the middle and connects to
the other, former weakly connected node. A resync will begin from
B to C.

A    B -> C

In the moment A connects to C the resync needs to be aborted:

A    B -- C
 `------>´

In case A demots, B & C do a handshake AFTER_UNSTABLE. In that
handshare the UUID comparis shows a split-brain. That can be
automatically resolved by syncing from the D_UP_TO_DATE node
to the D_OUTDATED node.

A    B <- C
 `-------´