drbd: Automatically resolve split-brain in a specific case with quorum
Consider three nodes A, B and C, with quorum enabled. A service
runs on node B at the moment:

   A
  / \
 C---B*

First action A gets isolated from the cluster

   A

 C---B*

A lost quorum, but that does not affect anyone. Now, the application is IO
idle at the moment, so no new current-UUID is generated at B at this
point. The next action is that B gets isolted

   A

 C   B*

Immediately after becoming isolated, the application still running on B
issues a write request. B creates a new UUID, sends that to C, and sends
the write. Since no ack comes back from C B realizes that it lost quorum.
The write request will be either completed with an error or forzen
(depending on the configured quorum mode). The application terminates after
seeing the IO error.
In the next step A gets released from isolation

   A*
  /
 C   B

C and A connect. No resync as they are still on the same datase. The
two nodes have quorum, the application gets started on A. It isseus as
write request, a new current UUID is created on A. With the next
step B gets connects to A first and then to B.

   A*
  / \
 C---B

B created a new current data UUID while it was isolated but DRBD never
completed an IO within that data epoch. B also crated a new current
data UUID and potentially completed IO requests, it has quorum.

This is a split-brain between B and A.

This commit introduces automatically resolving of such split-brain
situation. The condition is that one of the two partitions has quorum
and the other partition does not have quorum.