drbd: Propagate the current-UUID to diskless/secondary nodes
in a cluster with two (or more) diskless nodes, the exposed-uuid was not
propagated from the diskless primary to other diskless (secondary)
nodes if there is no direct connection between them.
Note: This is exactly how drbdmanage/linstor configure diskless nodes.

Now, if the primary role is moved gracefully between the diskless
nodes, the new primary does not have the correct exposed uuid.
This by itself does not create any harm. Most of the time will stay
completely unnoticed. When due to the first write a new current UUID
is generated on the diskless node the diskfull nodes adopts that
and the issue gets masked.

The issue only manifests itself it when the diskfull nodes get disconnected
(forcefully or by network disruption) and try to reconnect to the
diskless primary. They Outdate themselves as they suddenly see the
disagreement about the UUID of the up-to-date data.

To our support-force this was known as the "everything Outdated" bug.

Fix that by updating the exposed-uuid on diskless nodes while they
are secondary.