coccinelle/spatch based compat infrastructure
The idea is to shrink or completely remove drbd_wrappers.h and
instead patch the source code to be compatible with the kernel
we want to build for.

The patches are autogenerated using coccinelle semantic patches.

We want to distribute the conventional patches with any release
tar.gz's since we have to assume that older distros do not
come with spatch.

While building the necessary patches are applied to the source.
There is a "unpatch" make target to remove them in order to
use "git commit -a".