drbd: Downgrade disk if attached on secondary of two diskless nodes
Consider two diskless nodes, one primary. The secondary node now
attaches a disk.

While attaching it checks if there is a mismatch with the exposed
UUID of diskless peers. If yes, make sure the freshly attached
data-set is considered D_CONSISTENT at max. That disallows the
diskless primary to read/write from/to that.

If other nodes with the corrent UUID join later, that will
clean it up.