drbd: Make consistent disk state transitions for diskless/primary diskfull
In case a node with local disk connects to a primary/diskless and the
UUIDs do not match, ensure to present the node with disk as not
up to date. Presenting it D_CONSISTENT is sufficient. That has the
advantage that the whole cluster might be able to recover when
more connections get established.

While with setting it to D_OUTDATED, as it was done with patch
2b1959962c3f23520b40fe9541894f29996ae14f
automatic recovery becomes impossible.