compat: fixup KVER guessing from prepared kernel (-headers, -devel) tree
If we point our Makefile to some linux-headers-xyz or kernel-devel-qrl,
using KDIR=/usr/src/linux-headers-* or KDIR=/usr/src/kernels/*,
we then try to "guess" the kernel version by calling
make -s -C $(KDIR) kernelrelease ; which may fail if run as normal user,
depending on write permissions in various places, while still producing
the correct output.
With the addition of the "RETPOLINE" check and $(error ...) in the
linux Makefile, the correct output is missing, and this became relevant.

Try to use "kernelversion" first (which only uses "cached" information),
and try to fall back to "kernelrelease" only if the former fails.