drbd: close a small time window in which a node rejects to promote
When a secondary node looses contact to a primary and is D_UP_TO_DATE
in that moment it will voluntarily go to D_CONSISTENT for a short
time. It then does a two-phase commit to check if there is a different
path to that primary. If yes, it outdates itself. If no, it goes
back to D_UP_TO_DATE.

If in that small time window a promote request came in, it failed
since the disk was not D_UP_TO_DATE.

Fix that by waiting under this specific condition within the
drbd_set_role() function until the disk state is no longer D_CONSISTENT.