fixup 5be8c49eb8c9: drbd Consider resync data when calculating data in flight for pulling AHEAD
atomic_sub() accouting on rs_in_flight happens in got_BlockAck,
regardless of whether this was an "all-zeroes, P_RS_DEALLOCATED",
or a "normal P_RS_DATA_REPLY".

For now, we need to unconditionally do the atomic_add() accounting.
Todo: distinguish somehow in the ACK, maybe introduce a new packet,
or a new ID_SYNCER_ZEROES (-2ULL) or something like that,
but that needs a protocol bump.