drbd: introduce __tl_walk(,COMPLETION_RESUMED)
Fixes potential "hung" IO after a suspend-io/resume-io cycle.

Add a transfer-log walk variant that can pass in NULL as connection,
and use it to "wake up" all requests that are flagged as
RQ_COMPLETION_SUSP. If we don't do that, and there are no more pending
"events" for that particular request, we could potentially "forget" to
complete that request.

TODO in some followup changes:

 - for adminitrative "suspend-io" request, don't suspend completion
   in the first place, but let pending IO "drain".

 - double check that we resume *all* completions during recovery
   from the various failure-scenario triggered suspend-io reasons