drbd: anticipatory test_and_set() instead of test_bit() and set_bit()
Right now adm_del_minor() is still serialized on resource->adm_mutex.
If we lift that one day the test_and_set becomes important.
Having it that way easens maintenance for me, because I do not need to
look for the surounding mutex when I see these lines of code.