drbd: convert write_lock() to read_lock() around the UNREGISTERED bit in device->flags
in drbd_nl.c:
The state_rwlock is used here to make sure that it is still diskless while
the UNREGISTERED bits get set.

in drbd_main.c:
Again, it is used to make sure no state change happens while the
counts are increased.

It has nothing todo with the UNREGISTERED bit itself.