drbd: we should not unconditionally lc_unlock()
Even though it does no real harm (that I can see right now),
because the critical path is serialized with spin_lock(&device->al_lock),
we should strictly pair lc_unlock() only with a successful lock.

this is a fixup for
10b499cfb89a drbd: avoid potential deadlock during handshake
which moved the lock inside an if, but did not move the unlock.