drbd: Minor fix to UUID handling for dual primary/diskless and early writes
If we have two diskless nodes and two nodes with disk. Let's assume both
diskless nodes get promoted to primary, but have do not immediately execute
a write.  Now after both are settled being primary, one executes a write and
generates a new current UUID with that. That new UUID is accepted by the two
storage nodes but not by the other primary diskless node!

It seems that when doing live migration with KVM it closes the device
first on the source system before writing to the device on the
target device. So looks like nearly nobody was affeced by this bug.