drbd: expose openers via debugfs
So far we just wrote openers to the syslog, but that produces too much
spam. Now openers are exposed via:
/sys/kernel/debug/drbd/resources/$res/volumes/$volid/openers

We keep a list in the device. If the open counts drop to 0, we prune the
whole list. Otherwise we remove the oldest/first entry matching the
closing pid.

This should be as good as it gets, but it is not perfect. On one hand
the openening process might not be the closing one (mount vs. umount).
But we can assume that in these scenarios the open/close count drops to
0 and we remove all entries anyways. It could also happen that one
process (with the same pid) opens the device multiple times, and closes
it in reverse order. In such a case we would delete the first/oldest
entry, while we should have deleted the second one.