drbd: optimize the IO submit code path
instead of iterating over all objects of a resource to find out if
IO is suspended, cache the result of that with every state change
and only look at the cached result when submitting IO. This is valid
since all inputs can only be changed with a state change.

on my mid-range Skylake system Xeon Silver 4112 CPU @ 2.60GHz
this boost me from 360k read IOPs to 450k read IOPs.