Revert "drbd: fix "Logic BUG" after P_NEG_DREPLY"
This reverts commit 536f9b58c6f22e84ea3c7f45c084883b750adeb1.

When a secondary gets an IO error from its backend it sends back a P_NEG_ACK
and counts the op in the epoch. I.e. we need to keep it until we get the
P_BARRIER_ACK covering it.  Before this commit it aborts the connection with
"BAD! BarrierAck #2 received, but no epoch in tl!?"

BTW, this brings the code in sync with drbd-8.4 again.  The commit message
of the reverted commit gives no info of the nature of the bug that was
inteded to get fixed.  So, I assume that the commint I am reverting was
simply wrong.