drbd: fix missing set-out-of-sync for < D_INCONSISTENT peer
This fixes missing resyncs (and thus data corruption)
in scenarios described below.

When changing the receiving side to use the activity log
not only if "the" Primary peer was diskless, but always,
we lost the "set-out-of-sync" bit to track changed blocks
on behalf of the (not permanently, but temporarily detached)
diskless peer.

Which means that we don't know what to resync
after a "detach/keep modifying/attach" cycle on a Primary
besides what is covered by the activity log there.
Which is likely why this went undetected for so long: to notice,
modifications have to actually go to an area that was "cold"
during the detach.

A different scenario makes it immediately obvious, though:
"secondary/detach/primary/keep modifying/attach",
in which case on attach, no activity log is applied,
the bitmap(s) "seem" to be clean,
and we don't do any resync at all.

Very bad, very embarrassing :-(