drbd: keep peer-ack timer running when one connection goes down
because it might be necessary to send queued peer-acks into other
connections.

This basically reverts commit a88d3038

The issue was found while creating the "peer-acks" test for the test-suite,
so not tirggered in the wiled.

The consequences of the bug are quite severe:
A peer that does not get an expected peer_ack keeps the AL entry open.
That in turn can block a resync operation forever.