drbd: Initialize unused bitmap slots to 0 when creating it initially
When a new resource is up-ed for the first time the bitmap gets resized
from 0 to the agreed size when the nodes connect for the first time.

In case they do an initial resync the bits will be set during the
handshake.
But if the meta-data was set to contain the same current-UUID it
is intended that they do not do an initial resync, that is also
covered by the code that compares the current-UUIDs. It clears the
when the current-UUIDs are equal.
The problem was if such a resource, that started without initial
resync gets an additional node added later. While the UUID
comparison code correctly concludes that a partial resync is
sufficient, all the bits were set by the initial resize.

Fix that by initializing the bitmap to all 0 in the initial
resize. Then it collects all modifications while the
resource runs, and has that ready when the late node joins.