drbd: fix regression in request timeout handling
Symptoms:
Frequent messages like
| We did not send a P_BARRIER for 80360848ms > ko-count (7) * timeout (60 * 0.1s); drbd kernel thread blocked?
| Remote failed to finish a request within 80360852ms > ko-count (7) * timeout (60 * 0.1s)
with ridiculous times (which effectively are the uptime)

By introducing CONFIG_DRBD_TIMING_STATS, and disabling it by default,
we deactivated the detailed request timing in 9.0.15.
But still relied on those now initialized-to-zero timestamps
in the request_timer_fn().

Go back to using jiffies there,
for this purpose its granularity is more than good enough.