drbd: flush not yet submitted buffers before becoming secondary
While we had already everything in place to flush it out of DBRD,
we where missing things that linger around in the page cache above
us. Better flush that _before_ becoming secondary!

Of course the only way to trigger this is by using auto-promote.
In particular I was using auto-promote on a diskless node to trigger
the bug that inspired this change.

[26783.745591] drbd r0: role( Secondary -> Primary )
[26783.764158] drbd r0: role( Primary -> Secondary )
[26783.883627] drbd r0/0 drbd9: sending new current UUID: F7BBB1B405EE375E