drbd: Fix slow resync after node was diskless primary
When a node gets a write as diskless it needs to create a
new current uuid. So, the write gets queued for the submitter
thread. In the code path to queue it the ap_actlog_cnt counter
got increased but in submit_fast_path() it gets only decreased
if it has a private_bio (= is not diskless).

That in consequence causes drbd_try_rs_begin_io()
( via drbd_rs_should_slow_down() and drbd_rs_c_min_rate_throttle() )
to slow down the resync to the minimum resync rate.

Fixing that counter inc/dec fixes the slow resync.