drbd: fix lifetime of "need to apply activity log" metadata flag
During "crash recovery", we need to apply the "hot" area described by
the activity log to the out-of-sync bitmap whenever we are not sure
that there where no requests in flight to the backend.

"crash", and "need to apply activity log" are indicated by the presence
of the MDF_PRIMARY_IND flag in the DRBD "superblock".

A regression in DRBD 9 (wrt DRBD 8.4):
we cleared this flag when in Secondary role, which is wrong as long as
there is still a Primary around or there are still requests in flight.
In certain multiple-failure scenarios involing hard crashes (like
power failures) this could have led to incomplete resyncs.

This change sets the MDF_PRIMARY_IND when "someone" goes R_PRIMARY,
and only clears this on "clean detach", which is: non-forced
administrative detach, or when all nodes are known and R_SECONDARY,
and we know no (peer-) requests are in flight.